+++ app/grails-app/taglib/de/laser/SemanticUiDropdownTagLib.groovy
@@ -1,5 +1,6 @@
 package de.laser
 
+import com.k_int.kbplus.Org
 import com.k_int.kbplus.auth.User
 
 // Semantic UI
@@ -57,6 +58,40 @@ class SemanticUiDropdownTagLib {
 
         out <<   body()
 
+    }
+
+    // <semui:signedDropdown name="xyz" noSelection="Bitte auswählen .." from="${orgList}" signedIds="${signedOrgIdList}" />
+    def signedDropdown = { attrs, body ->
+
+        String id = ''
+        if (attrs.name) {
+            id = ' id="' + attrs.name + '" name="' + attrs.name + '" '
+        }
+        out << '<select class="ui fluid labeled search dropdown' + id + '">'
+
+        if (attrs.noSelection) {
+            out << '<option value="">' + attrs.noSelection + '</div>'
+        }
+
+        attrs.from?.each { item ->
+            out << '<option value="' + (item.class.name + ':' + item.id) + '">'
+
+            if (item instanceof Org) {
+                out << item.name
+
+                if (item.shortname) {
+                    out << ' (' + item.shortname + ') '
+                }
+            } else {
+                out << item
+            }
+
+            if (attrs.signedIds?.contains(item.id)) {
+                out << '&nbsp; &#10004;'
+            }
+            out << '</option>'
+        }
+        out << '</select>'
     }
 
     // <semui:actionsDropdown params="${params}"  />