+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+0.13.1
+
+- fixed adding providers/agencies modal for subscriptions
+- fixed start page typo
+
 0.13
 
 - reworked and improved finance handling